src/requirements.txt
PLY>=3.4
PyYAML>=3.11
